Created Stored Procs
Gotta change the final stored procedure to support dates
Have to change placeholder when you encounter a "Date" object
Support for Datetimes and other internal objects.
Only two pages left, must finish admin page by friday of next week, still working on the internal scripting side for the deployments page.
UI is nearly done, could do with fixing the Ticket header on the sidebar for the landing page as it shifts on different sized windows. 
Should run on the TV at home and on the mac with a smaller screen. 
Must also update the powershell script to support the rest of the files and include a Invoke-Webrequest so that if a version of node is not
present one can be downloaded from the node website. 
Must also update all of the footprint numbers to the link that was provided based off of the link that charles provided.